Move alias set to its own file
